Auth: Move SelfHostedInstructions and LostPassword to separate files
